pyproject.toml
[project]
name = "aiseed-weather"
version = "0.1.0"
description = "Global weather visualization desktop app using ECMWF and Open-Meteo data."
authors = [{ name = "Yasuhiro", email = "noreply@aiseed.dev" }]
license = { text = "AGPL-3.0-or-later" }
readme = "README.md"
requires-python = ">=3.13"
keywords = ["weather", "ecmwf", "open-meteo", "flet", "visualization", "aiseed"]

# Primary install path is conda-forge via environment.yml.
# Flet is installed via pip even in dev (beta with frequent releases).
# This list exists for packaging metadata.
dependencies = [
    "flet>=0.85",
    "ecmwf-opendata>=0.3.20",
    "xarray>=2024.1.0",
    "numpy>=1.26",
    "platformdirs>=4.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[tool.flet]
org = "dev.aiseed"
company = "AIseed"
copyright = "Copyright (c) 2026 Yasuhiro"

[tool.flet.app]
path = "src/aiseed_weather"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aiseed_weather"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
